# Update: Improve map parsing time - more efficient player attachment lookup