--- a/game-core/src/main/java/games/strategy/triplea/attachments/AbstractTriggerAttachment.java
+++ b/game-core/src/main/java/games/strategy/triplea/attachments/AbstractTriggerAttachment.java
@@ -51,16 +51,11 @@ public abstract class AbstractTriggerAttachment extends AbstractConditionsAttach
    */
   public static CompositeChange triggerSetUsedForThisRound(final GamePlayer player) {
     final CompositeChange change = new CompositeChange();
-    for (final TriggerAttachment ta : TriggerAttachment.getTriggers(player, null)) {
-      if (ta.getUsedThisRound()) {
-        final int currentUses = ta.getUses();
-        if (currentUses > 0) {
-          change.add(
-              ChangeFactory.attachmentPropertyChange(
-                  ta, Integer.toString(currentUses - 1), "uses"));
-          change.add(ChangeFactory.attachmentPropertyChange(ta, false, "usedThisRound"));
-        }
-      }
+    for (final TriggerAttachment ta :
+        TriggerAttachment.getTriggers(player, ta -> ta.getUsedThisRound() && ta.getUses() > 0)) {
+      change.add(
+          ChangeFactory.attachmentPropertyChange(ta, Integer.toString(ta.getUses() - 1), "uses"));
+      change.add(ChangeFactory.attachmentPropertyChange(ta, false, "usedThisRound"));
     }
     return change;
   }