--- a/game-core/src/main/java/games/strategy/triplea/attachments/UserActionAttachment.java
+++ b/game-core/src/main/java/games/strategy/triplea/attachments/UserActionAttachment.java
@@ -59,7 +59,7 @@ public class UserActionAttachment extends AbstractUserActionAttachment {
     TriggerAttachment trigger = null;
     for (final GamePlayer player : getData().getPlayerList().getPlayers()) {
       final TriggerAttachment ta = (TriggerAttachment) player.getAttachment(s[0]);
-      if(ta != null) {
+      if (ta != null) {
         trigger = ta;
         break;
       }