@Cernelius 2020-09-09
I'll certainly test this once it is merged (sadly, I don't know how to get the exact times, so I just look at a clock for the time the game takes to be available), also with one or more games not available in the repository.